refactor(product): 비속어 외부 API 공통으로 묶고, 상품 도메인에 연결 시켜줌.
