+++ src/main/java/kitchenpos/products/infra/ProductProfanityChecker.java
@@ -1,0 +1,19 @@
+package kitchenpos.products.infra;
+
+import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.products.tobe.domain.ProfanityChecker;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ProductProfanityChecker implements ProfanityChecker {
+    private final PurgomalumClient purgomalumClient;
+
+    public ProductProfanityChecker(PurgomalumClient purgomalumClient) {
+        this.purgomalumClient = purgomalumClient;
+    }
+
+    @Override
+    public boolean containsProfanity(String text) {
+        return purgomalumClient.containsProfanity(text);
+    }
+}
