+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -6,8 +6,9 @@ import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.products.dto.ProductChangePriceRequest;
 import kitchenpos.products.dto.ProductCreateRequest;
 import kitchenpos.products.dto.ProductResponse;
-import kitchenpos.products.infra.PurgomalumClient;
+import kitchenpos.common.infra.PurgomalumClient;
 import kitchenpos.products.tobe.domain.Product;
+import kitchenpos.products.infra.ProductProfanityChecker;
 import kitchenpos.products.tobe.domain.ProductRepository;
 import kitchenpos.products.tobe.domain.ProfanityChecker;
 import org.junit.jupiter.api.BeforeEach;
@@ -48,7 +49,7 @@ class ProductServiceTest {
         productRepository = new InMemoryProductRepository();
         menuRepository = new InMemoryMenuRepository();
         purgomalumClient = new FakePurgomalumClient();
-        profanityChecker = new ProfanityChecker(purgomalumClient);
+        profanityChecker = new ProductProfanityChecker(purgomalumClient);
         menuServiceClient = new FakeMenuServiceClient(menuRepository);
         productService = new ProductService(productRepository, profanityChecker, applicationEventPublisher, menuServiceClient);
     }
